Bugfix for #77 HeroMasteryTier Hero Encoding
FourCC data is encoded in the replay file as a bit-packed 32-bit
integer but was incorrectly being parsed as a bit-packed byte
sequence by the decoder.

The difference is how the bits are shifted to overlay the data stream
when the bit offset is not zero. A 4 byte sequence will not overlay the
same as a byte-encoded 32-bit int, even if the bytes are identical.

The fix is to read the data as a 32-bit bitstream first (effectively an
int), then pack as an int into a 4-character string.